Create ex07.cpp
Quadratic equations are of the form
                          a * xe2 + b * x + c = 0
To solve these, one uses the quadratic formula:
                       x = -b +- sqrt((be2) -4ac) / 2a
There is a problem, though: if b2-4ac is less than zero, then it will fail. Write
a program that can calculate x for a quadratic equation. Create a function that
prints out the roots of a quadratic equation, given a, b, c. When the program
detects an equation with no real roots, have it print out a message. How do you
know that your results are plausible? Can you check that they are correct?