Chaper 6: create ex03.cpp
Add a factorial operator: use a suffix ! operator to represent “factorial.” For
example, the expression 7! means 7 * 6 * 5 * 4 * 3 * 2 * 1. Make ! bind tighter
than * and /; that is, 7*8! means 7*(8!) rather than (7*8)!. Begin by modifying
the grammar to account for a higher-level operator. To agree with the standard
mathematical definition of factorial, let 0! evaluate to 1. Hint: The calculator
functions deal with doubles, but factorial is defined only for ints, so just for
x!, assign the x to an int and calculate the factorial of that int.