Chapter 08: solve ex01.cpp
Modify the calculator program from chapter 7 to make the input stream an
explicit parameter (as shown in §8.5.8), rather than simply using cin. Also
give the Token_stream constructor (§7.8.2) an istream& parameter so that when
we figure out how to make our own istreams (e.g., attached to files, we can
use the calculator for those. Hint: Don't try to copy an istream.