Chapter 08: solve ex09.cpp
Write a function maxv() that returns the largest element of a vector argument
and also computes the mean and the median, Do not use global variables. Either
return a struct containing the results or pass them back through reference
arguments. Which of the two ways of returning several result values do you
prefer and why?