Chapter 09: solve ex02.cpp
Design and implement a Name_pairs class holding(name, age) where name is a
string and age is a double. Represent that as vector<string> (called name)
and a vector<double> (called age) member. Provide an input operation
read_names9) that reads a series of names. Provide a read_ages() operation
that prompts the user for an age for each name. Provide print() operation
that prints out the (name[i], age[i]) pairs (one per line) in the order
determined by the name vector. Provide sort() operation that sorts the name
vector in alphabetical order and reorganizes age vector to match. Implement
all "operations" as member functions. Test the class (of course: test early
and often).