Chapter 09: solve ex05.cpp
This exercise and the next few require you to design and implement a Book
class, such as you can imagine as part of a software for a library. Class book
should have members for the ISBN, title, author, and copyright date. Also
store data on whether or not the book is checked out. Create functions for
returning those data values. Create functions for checking a book in and out.
Do simple validation of data entered into a book; for example, accept ISBNs only
of the form n-n-n-x where n is an integer and x is a digit or a letter.
Store an ISBN as a string.