Chapter 09: solve ex14.cpp
Design and implement a Money class for calculations involving dollars and
cents where arithmetic has to be accurate to the last cent using the 4/5
rounding rule (.5 of a cent rounds up; anything less than .5 rounds down).
Represent a monetary amount as a number of cents in a long int, but input
and output as dollars and cents, e.g., $123.45. Do not worry about amounts
that dont't fit into a long int.