Chapter 10: implement exercise 10
Add a command from x to the calculator from Chapter 7 that makes it take input from a file x.
Add a command to y to the calculator that makes it write its output (both standard output and
error output) to file y. Write a collecion of test cases based on ideas from §7.3 and use that
to test the calculator. Disscuss how you would use these commands for testing.