Chapter 14: create drill 07
Define a class B2 with a pure virtual function pvf(). Define a class D21
with a string data member and a member function that overrides pvf();
D21::pvf() should output the value of the string. Define a class D22 that
is just like D21 except that its data member is an int. Define function f()
that takes a B2& argument and calls pvf() for its argument. Call f() with a
D21 and a D22.