Chapter 14: create ex05.cpp
Define a Stripped_rectangle where instead of fill, the rectangle is "filled"
by drawing one-pixel-wide horizontal lines across the inside of the rectangle
(say, draw every second line like that). You may have to play with the width
of lines and the line spacing to get a pattern you like.