Chapter 14: solve exercise 10
Define a class Pseudo_window that looks as much like a window as you
can make it without heroic efforts. It should have rounded corners,
a label, and control icons. Maybe you could add some fake "contents",
such as an image. It need not actually do anything. It is acceptable
(and indeed recommended) to have it appear within a Simple_window.