Chapter 14: solve exercise 15
Most class hierarchies have nothing to do with graphics. Define a class
Iterator with a pure virtual function next() that returns a double* (see
Chapter 17). Now derive Vector_iterator and List_iterator from Iterator
so that next() for Vector_iterator yields a pointer to the next element
of a vector<double> and List_iterator does the same for list<double>.
You initialize a Vector_iterator with a vector<double> and the first call
of next() yields a pointer to its first element, if any. If there is no
next element, return 0. Test this by using a function void print(Iterator&)
to print the elements of a vector<double> and list<double>.