Chapter 14: solve exercise 16
Define a class Controller with four virtual functions on(), off(), set_level(int),
and show(). Derive at least two classes from Controller. One should be a simple
test calss where show() prints out whether the class is set on or off and what
is the current level. The second derived class should somehow control the line color
of a Shape; the exact meaning of "level" is up to you to define. Try to find a third
"thing" to control with such a Controller class.